Fixed handshake bug in mbedTLS wrapper
Fixes a race condition when using mbedTLS: the TLS handshake would
abort if no data had yet arrived from the peer.

This happened because in the OpenSSL-to-mbedTLS wrapper code, the
'handshake' method (ssl_pm_handshake) behaved incorrectly when the BIO
has an incomplete read or write:
- It set ssl->err to an mbedTLS error code instead of an OpenSSL code
- It returned 0 instead of -1
This then caused SSL_get_error to return the wrong results, making it
look as though the handshake had aborted when it just needed to read
or write more data.

In addition, I fixed two places in the code where 'errno' was compared
with a literal 11 instead of EAGAIN. This is not compatible between
platforms, as errno values are numbered differently on different OSs
(on BSDs and Apple platforms, EAGAIN is 35!)