Added API to get/set mbedTLS credentials during callback
LWS_CALLBACK_OPENSSL_LOAD_EXTRA_SERVER_VERIFY_CERTS callback passes a
pointer to the SSL context. If using OpenSSL that context is a known
type that has APIs, but if using mbedTLS there is nothing exposed to
operate on it.
This commit adds functions to get and set X.509 certs and key-pairs
on that context, when using mbedTLS.